+++ serializers/resource.js
@@ -5,12 +5,6 @@ var JSONAPISerializer = require('jsonapi-serializer').Serializer;
 var Schemas = require('../generators/schemas');
 var logger = require('../services/logger');
 
-function toKebabCase(string) {
-  // NOTICE: Support for the collections beginning with an underscore.
-  if (string[0] === '_') { return '_' + _.kebabCase(string); }
-  return _.kebabCase(string);
-}
-
 function ResourceSerializer(Implementation, model, records, integrator,
   opts, meta) {
   var schema = Schemas.schemas[Implementation.getModelName(model)];
@@ -54,7 +48,7 @@ function ResourceSerializer(Implementation, model, records, integrator,
           } else if (field.reference) {
             var referenceType = field.reference.split('.')[0];
             var referenceSchema = Schemas.schemas[referenceType];
-            typeForAttributes[field.field] = toKebabCase(referenceType);
+            typeForAttributes[field.field] = referenceType;
 
             if (!referenceSchema) {
               logger.error('Cannot find the \'' + referenceType +
@@ -120,8 +114,7 @@ function ResourceSerializer(Implementation, model, records, integrator,
       formatDateonly(records);
     }
 
-    var typeName = toKebabCase(schema.name);
-    return new JSONAPISerializer(typeName, records, serializationOptions);
+    return new JSONAPISerializer(schema.name, records, serializationOptions);
   };
 }